+++ src/LinqInfer/Learning/Features/IValueConverter.cs
@@ -8,6 +8,6 @@ namespace LinqInfer.Learning.Features
     public interface IValueConverter
     {
         bool CanConvert(Type type);
-        float Convert(object value);
+        double Convert(object value);
     }
 }
